renamed controllers to grabandthrow, also added grabandthrow to main menu
