+++ BasicSample/UserSettings/EditorUserSettings.asset
@@ -12,28 +12,28 @@ EditorUserSettings:
       value: 224247031146467a3f230931186c7e223b120b22623d28393930
       flags: 0
     RecentlyUsedScenePath-2:
-      value: 22424703114646760c07020f152d4a5e0319113e35
-      flags: 0
-    RecentlyUsedScenePath-3:
-      value: 224247031146467a3f3e00231e2610312427142b222d68252320092a
-      flags: 0
-    RecentlyUsedScenePath-4:
       value: 22424703114646730c00081602225c1b1f191f6504292834193b1c30e9e93a3facf238e0f323
       flags: 0
-    RecentlyUsedScenePath-5:
-      value: 224247031146467c1f0f0e031e276b1804180f650b3a27320c271907eaf23b2facf238e0f323
-      flags: 0
-    RecentlyUsedScenePath-6:
+    RecentlyUsedScenePath-3:
       value: 2242470311464672031a093011204b1919195703223c23222c2a093aedee7a2decee22f0
       flags: 0
-    RecentlyUsedScenePath-7:
+    RecentlyUsedScenePath-4:
+      value: 224247031146467a3f3e00231e2610312427142b222d68252320092a
+      flags: 0
+    RecentlyUsedScenePath-5:
       value: 224247031146467a3f3c0d3b13224c0459362a182d3125313e3d5326ece92021
+      flags: 0
+    RecentlyUsedScenePath-6:
+      value: 2242470311464678020018301f2f5315040457092326322222251136f0f37a2decee22f0
+      flags: 0
+    RecentlyUsedScenePath-7:
+      value: 22424703114646760c07020f152d4a5e0319113e35
       flags: 0
     RecentlyUsedScenePath-8:
       value: 224247031146467a3f2f0221182c4d5f372539242f202922633c133af6f9
       flags: 0
     RecentlyUsedScenePath-9:
-      value: 2242470311464678020018301f2f5315040457092326322222251136f0f37a2decee22f0
+      value: 224247031146467c1f0f0e031e276b1804180f650b3a27320c271907eaf23b2facf238e0f323
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
